[Refact] panel 이벤트 수정
- addButtonEvnets가 실행될 때 같이 실행될 수 있도록 했습니다.
